x-pack/filebeat/input/httpjson: improve code locality of chain helpers
